refactor: extract editor document lifecycle
